move hash partial worker into new file
